build: [D4·D8·D12] iOS 빌드 로그 삼킴 수정 + Android CI 신설 + 컴파일 경고 0건
- build-ipa.sh: xcodebuild 를 `grep -E "warning: [^d]" || true` 에 파이프해
  ① `warning: deprecated…` 가 d 로 시작한다는 이유로 통째로 필터링되고
  ② `|| true` 가 실패 종료코드를 삼켰다. → 로그 파일 리다이렉트 + 실제 exit code +
  "ARCHIVE SUCCEEDED" 문자열 이중 검증. (Android 릴리즈를 깨뜨렸던 것과 같은 패턴)
- GL_Shared android 타겟에 withHostTest{} 추가 — commonTest 3건(KeyStatRules·Format·Md5)이
  어떤 타깃에서도 실행되지 않고 있었다. 이제 :GL_Shared:testAndroidHostTest 로 실행.
- .github/workflows/android-ci.yml 신설: PR·main push 마다 shared 테스트 +
  compileReleaseKotlin + assembleDebug + lint. (릴리즈 컴파일 실패가 main 에 머지된 이력 재발 방지)
- 컴파일 경고 19 → 0: -Xexpect-actual-classes(12건) + deprecated LocalClipboardManager →
  LocalClipboard(1건) + 잉여 else/변환/안전호출(6건).

+++ GL_Shared/src/commonMain/kotlin/com/gatcha/log/util/Md5.kt
@@ -15,10 +15,10 @@ fun md5Hex(input: String): String {
         ByteArray(8) { i -> ((origLenBits ushr (8 * i)) and 0xFF).toByte() }
 
     // 초기 상태
-    var a0 = 0x67452301.toInt()
+    var a0 = 0x67452301
     var b0 = 0xefcdab89.toInt()
     var c0 = 0x98badcfe.toInt()
-    var d0 = 0x10325476.toInt()
+    var d0 = 0x10325476
 
     val s = intArrayOf(
         7, 12, 17, 22, 7, 12, 17, 22, 7, 12, 17, 22, 7, 12, 17, 22,
